+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - warning about the `projects at` value when using default config file
+- support for NO_COLOR and FORCE_COLOR env vars
 
 ### Changed